Convert polynomial to integer, restore some linearity, add comments
This uses |amethyst's version after Horner's rule from comments on the
PR.  This is nearly the same polynomial as before, but slightly better
fit.

The calculation in this version restores the old behavior of being
linear after 45 (adjusted slightly for this polynomial), which I didn't
understand when I made my first commit. Dropping this part will pull
down spell failure rates noticeably in the higher ranges with this
polynomial, and may actually be worth considering -- based on my testing
I don't think it'll drastically affect casting for the meaningful part
of the fail range, and it is simpler (removes another breakpoint).