Add scale parameters, for better spell power resolution
This adds a scale parameter to `stepdown_spellpower` and to
`calc_spell_power`, and uses the scale parameter to get higher
resolution in `raw_spell_fail`.  This results in much more continuous
behavior of spell fail rates when training, though it is still not high
enough resolution to be perfectly continuous, especially at high skill
levels.

The behavior of the two changed functions with default arguments remains
the same.